Update actions footer listview locations
