+++ client/packages/system/src/Location/ListView/ListView.tsx
@@ -14,6 +14,7 @@ import { useLocation, LocationRowFragment } from '../api';
 import { AppBarButtons } from './AppBarButtons';
 import { LocationEditModal } from './LocationEditModal';
 import { Toolbar } from './Toolbar';
+import { Footer } from './Footer';
 
 const LocationListComponent: FC = () => {
   const {
@@ -73,7 +74,7 @@ const LocationListComponent: FC = () => {
           location={entity}
         />
       )}
-      <Toolbar data={locations} filter={filter} />
+      <Toolbar filter={filter} />
       <AppBarButtons onCreate={() => onOpen()} />
       <DataTable
         id="location-list"
@@ -91,6 +92,7 @@ const LocationListComponent: FC = () => {
           />
         }
       />
+      <Footer data={locations} />
     </>
   );
 };
